+++ ReleaseNotes.md
@@ -15,6 +15,9 @@ Renamed functions
 - _getMapCenter() removed. Use getMapCenter() instead.
 - _updateDomainImpl() removed. Use updateDomain() instead.
 - panMapToLonLat() removed. Use centerMap() instead.
+- panMapEast()/panMapWest()/panMapNorth()/panMapSouth() removed. Use panMapByPixels() instead.
+- zoomIn()/zoomOut() removed. Use adjustZoomLevel() instead.
+- zoomTo() removed. Use setZoomLevel() instead.
 
 Unused functions removed:
   - _ensureExists()
@@ -30,6 +33,7 @@ Unused functions removed:
 
 Added functions so internal references don't need to be called:
 - getMapElementId()
+- getCurrentExtent()
 
 
 ### mapping/mapmodule/plugin/vectorlayer